Initial shot at scatter gather support.
Adds APIs to add a nested buffer structure to be
scatter gathered. Uses the new BC_TRANSACTION_SG
and BC_REPLY_SG commands in the kernel driver to
make sure libbinder is fully backwards compatible.

This is likely not the final API, and a lot of
checking is still missing from this implementation.
Change-Id: I66bd11907cb0697aeffa4b710efa4aa75826e9c4