Pad binder_fd_array_object.
binder_fd_array_object starts with a 4-byte header,
followed by a few 8-byte fields. When compiling this
for 32-bit, the 8-byte fields can be 4-byte aligned;
on 64-bit, the 8-byte fields must be 8-byte aligned.
This can cause issues when combining a 64-bit kernel
with a 32-bit userspace, as fugu does. Pad the struct
to make sure the 8-byte fields start at an 8-byte
boundary.

Bug: 34635658
Test: hidl_test
Change-Id: I8108b894cc430b359134b632cf07eb90e616458f