Test kernel scheduler behaviors with libhwbinder
It tests
1. supports priority inheritance.
2. utilize the sync flag.

It also makes statistic on average, best and worst latency
as well as number of deadline miss with a configurable
deadline specified in us. The output is in json format

usage:./libhwbinder_latency \
  -i 1              # number of iterations
  -pair 4           # number of process pairs
  -deadline_us 2500 # deadline in us
  -v                # debug
  -trace            # halt the trace on a dealine hit

This -trace option stops trace once it detects a transaction
duration over the deadline. By writing '0' to
/sys/kernel/debug/tracing and halt the process. The tracelog is
then available on /sys/kernel/debug/trace

Bug: 36705188
Test: sailfish with prebuilt kernel on oc-dev
Change-Id: I702222993a95f7dde0dfb9f75724766f9924a05b
Merged-In: If96268915ee4b602775b8840786f6ba477895427