Add support for min scheduling policy to node.
The binder driver has always supported setting a
min nice value for transactions. It looks like this
was always set to an out-of-range value (0x7f);
correct this to use a value of 0 by default (which
corresponds to nice 0).

Additionally, the new driver supports setting
a minimum scheduler *policy* as well, by reserving
2 bits for the scheduler policy in the flat_binder_object
structure. Add APIs to set this policy on a server object,
and set them properly when passing the object into
the driver.

Bug: 37293077
Bug: 63899698
Test: verified /d/binder output
Merged-In: I1206320c37e89efd523e342b5d56179e403eac50
Change-Id: I1206320c37e89efd523e342b5d56179e403eac50