Bring back starvation messages for single thread.
These kind of messages use to cause a lot of confusion, see b/37647467
binder thread pool (1 threads) starved for 4922 ms

However it is useful in some cases to find out some HAL is taking too
much time causing framework to freeze. (b/74187627)
And this log was good signal for bugs in the system in the past, usually
the scheduler but not always.

Bug: 74436227
Test: mma
Change-Id: I4706918d0938ec2eeba50429fa734a3b6253e35b