ProcessState: don't require /dev/hwbinder on host.
On host, we allow ProcessState to be created (and for subsequent calls
to it to fail) in order to allow unit tests to run on host which
frequently open up ProcessState. For instance, this happens when Parcel
objects are constructed in certain cases.

This isn't ideal, but it is the workaround right now we have to avoid
needing huge deltas between device/host code.

Fixes: 145570461
Test: ProcessState::self() on host doesn't crash.
Change-Id: Ie71ef1659ac8f283f562c289bd978835251d5aba