Disallow shrinking threadpool size once started.
If a larger threadpool is started, and concurrent binder transactions
are cause these threads to actually be started, and then we shrink
the threadpool, these extra threads won't be shut down. So, when we
detect that this race may happen, we abort with an error.

It's best practice to setup a threadpool at process startup time.

Bug: 147699567
Test: TH
Change-Id: Ideb091efdb845e9c286788fcd6a7e23e65ebe9df